Correct Geometry::Object test so that one of its helper functions doesn't return a local stack object by value. It now uses shared pointers instead. Also clear a warning. Re #1001

+++ Code/Mantid/Geometry/inc/MantidGeometry/Objects/Track.h
@@ -123,7 +123,7 @@ class DLLExport Track
 
   LType::const_iterator begin() const { return Link.begin(); }   ///< Iterator pointing to start of collection
   LType::const_iterator end() const { return Link.end(); }       ///< Iterator pointing one-past-the-end of collection
-  int count() const { return Link.size(); }       ///< The number of link items in the track
+  int count() const { return static_cast<int>(Link.size()); }       ///< The number of link items in the track
 
   // tests
   int nonComplete() const;
